requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
gunicorn==23.0.0
tensorflow-cpu==2.15.0
numpy==1.26.4
Pillow==10.4.0
PyYAML==6.0.1